requirements.txt
streamlit
pytrends
pandas
Pillow